+++ packages/cinterop/src/darwin/kotlin/io/realm/internal/interop/Errors.kt
@@ -62,6 +62,7 @@ fun coreErrorAsThrowable(nativeValue: realm_errno, message: String?): RealmCoreE
         realm_errno.RLM_ERR_FILE_ACCESS_ERROR -> RealmCoreFileAccessErrorException(message)
         realm_errno.RLM_ERR_FILE_PERMISSION_DENIED -> RealmCoreFilePermissionDeniedException(message)
         realm_errno.RLM_ERR_DELETE_OPENED_REALM -> RealmCoreDeleteOpenRealmException(message)
+        realm_errno.RLM_ERR_ILLEGAL_OPERATION -> RealmCoreIllegalOperationException(message)
         else -> RealmCoreUnknownException(message)
     }
 }