+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/IrUtils.kt
@@ -86,6 +86,7 @@ import org.jetbrains.kotlin.name.Name
 import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
 import org.jetbrains.kotlin.resolve.descriptorUtil.getSuperInterfaces
 import java.util.function.Predicate
+import java.util.function.Predicate
 import org.jetbrains.kotlin.resolve.descriptorUtil.isSubclassOf
 import org.jetbrains.kotlin.resolve.descriptorUtil.module