+++ packages/library-sync/src/commonMain/kotlin/io/realm/mongodb/EmailPasswordAuth.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2022 Realm Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.realm.mongodb
 
 import io.realm.mongodb.exceptions.AppException
@@ -15,9 +31,8 @@ public interface EmailPasswordAuth {
      * 6 and 128 characters long.
      * @throws io.realm.mongodb.exceptions.UserAlreadyExistsException if this email was already
      * registered.
-     * @throws io.realm.mongodb.exceptions.ServiceException All API's that talk to Atlas App
-     * Services through a HTTP request can fail in a variety of ways. See [AppException] for details
-     * about the specialized subclasses.
+     * @throws io.realm.mongodb.exceptions.ServiceException for other failures that can happen when
+     * communicating with App Services. See [AppException] for details.
      */
     public suspend fun registerUser(email: String, password: String)
 
@@ -28,9 +43,8 @@ public interface EmailPasswordAuth {
      * @param tokenId the id of the confirmation token.
      * @throws io.realm.mongodb.exceptions.UserAlreadyConfirmedException if this email was already
      * confirmed.
-     * @throws io.realm.mongodb.exceptions.ServiceException All API's that talk to Atlas App
-     * Services through a HTTP request can fail in a variety of ways. See [AppException] for details
-     * about the specialized subclasses.
+     * @throws io.realm.mongodb.exceptions.ServiceException for other failures that can happen when
+     * communicating with App Services. See [AppException] for details.
      */
     public suspend fun confirmUser(token: String, tokenId: String)
 
@@ -42,9 +56,8 @@ public interface EmailPasswordAuth {
      * this email.
      * @throws io.realm.mongodb.exceptions.UserAlreadyConfirmedException if the user was already
      * confirmed.
-     * @throws io.realm.mongodb.exceptions.ServiceException All API's that talk to Atlas App
-     * Services through a HTTP request can fail in a variety of ways. See [AppException] for details
-     * about the specialized subclasses.
+     * @throws io.realm.mongodb.exceptions.ServiceException for other failures that can happen when
+     * communicating with App Services. See [AppException] for details.
      */
     public suspend fun resendConfirmationEmail(email: String)
 
@@ -56,9 +69,8 @@ public interface EmailPasswordAuth {
      * this email.
      * @throws io.realm.mongodb.exceptions.UserAlreadyConfirmedException if the user was already
      * confirmed.
-     * @throws io.realm.mongodb.exceptions.ServiceException All API's that talk to Atlas App
-     * Services through a HTTP request can fail in a variety of ways. See [AppException] for details
-     * about the specialized subclasses.
+     * @throws io.realm.mongodb.exceptions.ServiceException for other failures that can happen when
+     * communicating with App Services. See [AppException] for details.
      */
     public suspend fun retryCustomConfirmation(email: String)
 
@@ -68,9 +80,8 @@ public interface EmailPasswordAuth {
      * @param email the email of the user.
      * @throws io.realm.mongodb.exceptions.UserNotFoundException if no user was registered with
      * this email.
-     * @throws io.realm.mongodb.exceptions.ServiceException All API's that talk to Atlas App
-     * Services through a HTTP request can fail in a variety of ways. See [AppException] for details
-     * about the specialized subclasses.
+     * @throws io.realm.mongodb.exceptions.ServiceException for other failures that can happen when
+     * communicating with App Services. See [AppException] for details.
      */
     public suspend fun sendResetPasswordEmail(email: String)
 
@@ -98,9 +109,8 @@ public interface EmailPasswordAuth {
      * existing user.
      * @throws io.realm.mongodb.exceptions.BadRequestException if the input tokens where
      * rejected by the server for being malformed.
-     * @throws io.realm.mongodb.exceptions.ServiceException All API's that talk to Atlas App
-     * Services through a HTTP request can fail in a variety of ways. See [AppException] for details
-     * about the specialized subclasses.
+     * @throws io.realm.mongodb.exceptions.ServiceException for other failures that can happen when
+     * communicating with App Services. See [AppException] for details.
      */
     public suspend fun resetPassword(token: String, tokenId: String, newPassword: String)
 }