+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [Sync] Support for the extension property `Realm.syncSession`, which returns the sync session associated with the realm.
 * [Sync] Support for `SyncSession.downloadAllServerChanges()` and `SyncSession.uploadAllLocalChanges()`.
 * [Sync] Support for `App.allUsers()`.
+* [Sync] Support for `App.switchUser(user)`.
 * [Sync] Support for `User.remove()`.
 
 ### Fixed