+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [Sync] Support for `SyncSession.downloadAllServerChanges()` and `SyncSession.uploadAllLocalChanges()`.
 * [Sync] Support for `App.allUsers()`.
 * [Sync] Support for `App.switchUser(user)`.
+* [Sync] Support for `App.switchUser(user)`.
 * [Sync] Support for `User.remove()`.
 
 ### Fixed